requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0
